add DELETE /api/location/:id

+++ server/models/department.modals.js
@@ -24,23 +24,31 @@ module.exports.create = (obj, cb) => {
 }
 
 module.exports.get_by_id = (id, cb) => {
+    if (!id) {
+        throw new Error("Id is required");
+    }
+
     const db_connect = db.getDb();
 
     db_connect.collection("departments").findOne({ _id: ObjectId(id) }, cb)
 };
 
 module.exports.delete = (id, cb) => {
+    if (!id) {
+        throw new Error("Id is required");
+    }
+
     const db_connect = db.getDb();
 
     db_connect.collection("departments").deleteOne({ _id: ObjectId(id) }, cb);
 }
 
 module.exports.update = (id, obj, cb) => {
-    const db_connect = db.getDb();
-
     if (!obj.name) {
         throw new Error("Name field is required");
     }
+
+    const db_connect = db.getDb();
 
     db_connect.collection("departments").updateOne({ _id: ObjectId(id) }, {
         $set: { name: obj.name }
