+++ server/models/personnel.modals.js
@@ -30,18 +30,30 @@ module.exports.get_all = (cb) => {
 }
 
 module.exports.get_by_id = (id, cb) => {
+    if (!id) {
+        throw new Error("Id is required");
+    }
+
     const db_connect = db.getDb();
 
     db_connect.collection("personnel").findOne({ _id: ObjectId(id) }, cb)
 }
 
 module.exports.delete = (id, cb) => {
+    if (!id) {
+        throw new Error("Id is required");
+    }
+
     const db_connect = db.getDb();
 
     db_connect.collection("personnel").deleteOne({ _id: ObjectId(id) }, cb);
 }
 
 module.exports.update_personnel = (id, body, cb) => {
+    if (!id) {
+        throw new Error("Id is required");
+    }
+
     const db_connect = db.getDb();
 
     let fields = {};
